Reset zokrates_bellman back to state in develop branch
