+++ zokrates_bellman/src/groth16.rs
@@ -10,7 +10,7 @@ use zokrates_proof_systems::{Backend, MpcBackend, NonUniversalBackend, Proof, Se
 
 use crate::Bellman;
 use crate::Computation;
-use crate::{parse_g1, parse_g2, serialization};
+use crate::{parse_g1, parse_g2};
 use phase2::MPCParameters;
 use rand_0_4::Rng;
 use std::io::{Read, Write};
@@ -93,7 +93,7 @@ impl<T: Field + BellmanFieldExtensions> NonUniversalBackend<T, G16> for Bellman 
         let mut pk: Vec<u8> = Vec::new();
         parameters.write(&mut pk).unwrap();
 
-        let vk = parameters_to_verification_key::<T>(&parameters);
+        let vk = serialization::parameters_to_verification_key::<T>(&parameters);
         SetupKeypair::new(vk, pk)
     }
 }
@@ -153,20 +153,47 @@ impl<T: Field + BellmanFieldExtensions> MpcBackend<T, G16> for Bellman {
     }
 }
 
-pub fn parameters_to_verification_key<T: Field + BellmanFieldExtensions>(
-    parameters: &Parameters<T::BellmanEngine>,
-) -> VerificationKey<G1Affine, G2Affine> {
-    VerificationKey {
-        alpha: parse_g1::<T>(&parameters.vk.alpha_g1),
-        beta: parse_g2::<T>(&parameters.vk.beta_g2),
-        gamma: parse_g2::<T>(&parameters.vk.gamma_g2),
-        delta: parse_g2::<T>(&parameters.vk.delta_g2),
-        gamma_abc: parameters
-            .vk
-            .ic
-            .iter()
-            .map(|g1| parse_g1::<T>(g1))
-            .collect(),
+pub mod serialization {
+    use super::*;
+    use pairing::from_hex;
+    use zokrates_proof_systems::{G1Affine, G2Affine};
+
+    pub fn parameters_to_verification_key<T: Field + BellmanFieldExtensions>(
+        parameters: &Parameters<T::BellmanEngine>,
+    ) -> VerificationKey<G1Affine, G2Affine> {
+        VerificationKey {
+            alpha: parse_g1::<T>(&parameters.vk.alpha_g1),
+            beta: parse_g2::<T>(&parameters.vk.beta_g2),
+            gamma: parse_g2::<T>(&parameters.vk.gamma_g2),
+            delta: parse_g2::<T>(&parameters.vk.delta_g2),
+            gamma_abc: parameters
+                .vk
+                .ic
+                .iter()
+                .map(|g1| parse_g1::<T>(g1))
+                .collect(),
+        }
+    }
+
+    pub fn to_g1<T: BellmanFieldExtensions>(
+        g1: G1Affine,
+    ) -> <T::BellmanEngine as Engine>::G1Affine {
+        <T::BellmanEngine as Engine>::G1Affine::from_xy_unchecked(
+            from_hex(&g1.0).unwrap(),
+            from_hex(&g1.1).unwrap(),
+        )
+    }
+    pub fn to_g2<T: BellmanFieldExtensions>(
+        g2: G2Affine,
+    ) -> <T::BellmanEngine as Engine>::G2Affine {
+        match g2 {
+            G2Affine::Fq2(g2) => {
+                let x = T::new_fq2(&(g2.0).0, &(g2.0).1);
+                let y = T::new_fq2(&(g2.1).0, &(g2.1).1);
+                <T::BellmanEngine as Engine>::G2Affine::from_xy_unchecked(x, y)
+            }
+            _ => unreachable!(),
+        }
     }
 }
